requirement.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
distro==1.9.0
exceptiongroup==1.3.0
gitdb==4.0.12
GitPython==3.1.40
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.2
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.98.0
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.4
smmap==5.0.2
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
